refactor(R-01-7.2): delete legacy RepoRowView.swift, migrate SemanticScoreBadge
R-01 §3.1.5 Step 7.2 部分完成：

- 删除 Features/Home/RepoRowView.swift（含 RepoRowView / RepoRowCard / LegacyRepoRowSurface），共 -245 行
- 将仍被 UnifiedRepoRow 使用的 SemanticScoreBadge 迁移到 Shared/Components/UnifiedRepoRow.swift（唯一调用方）

Trending/Weekly 相关旧文件（TrendingRepoRowView / RepoDetailView Trending 分支 /
WeeklyDetailView star/unstar / TrendingViewModel.subscribedRepoIDs）暂留，
等 6.2/6.3 后端 envelope 上线后随场景接入 PR 一起删除。

xcodegen + 360/360 tests pass.
