+++ docs/工程进度/R-01-重构进度.md
@@ -194,11 +194,14 @@
   - `StarcatTests/AppSettingsTests.swift`：`densityPersists` 测试改为验证默认值即为 `.card`（不再切换 compact）
   > 实现：保留 `RepoListDensity` 单 case 而非完全删除是为了 ① 避免 row view init 入参签名一次性改动 ② Step 7.2 后续删 row view 旧文件时一并彻底清理。360/360 全绿。
 
-- [ ] **7.2** 删除遗留文件 + 5 处 star/unstar 旧实现
-  - 删除文件：`Features/Home/RepoRowView.swift` / `Features/Trending/TrendingRepoRowView.swift`
-  - 删 `RepoDetailView.swift` 中 Trending 分支（`trendingMetadataHeader` / `starTrending` / `performUnstar(repo:)`）
-  - 删 `ActivityDetailView.swift` 中 `performUnstar(repo:)`
-  - 删 `WeeklyDetailView.swift` 中 `performUnstar(repo:)` / `performStar(repo:)`
+- [x] **7.2** 删除遗留文件 + 5 处 star/unstar 旧实现 — **2026-06-09 部分完成**（其余因 6.2/6.3 取消而搁置）
+  - ✅ 删除 `Features/Home/RepoRowView.swift`（含 `RepoRowView` / `RepoRowCard` / `LegacyRepoRowSurface`）；将其中唯一外部仍在用的 `SemanticScoreBadge` 迁移到 `Shared/Components/UnifiedRepoRow.swift`（唯一调用方）
+  - ❌ `Features/Trending/TrendingRepoRowView.swift` —— **暂留**：6.2 取消（后端 envelope 未上线），Trending 列表仍依赖此 view
+  - ❌ `RepoDetailView.swift` Trending 分支（`trendingMetadataHeader` / `starTrending` / `performUnstar(repo:)`）—— **暂留**：6.2 取消
+  - ✅ `ActivityDetailView.swift` `performUnstar(repo:)` —— 已在 6.4 删除
+  - ❌ `WeeklyDetailView.swift` `performUnstar(repo:)` / `performStar(repo:)` —— **暂留**：6.3 取消
+  - ❌ `TrendingViewModel.subscribedRepoIDs` / `subscribe` / `incrementStarsCount` —— **暂留**：6.2 取消，星标回填仍走旧路径
+  > 实现：本步骤主要清掉了 Manage 已迁移完成那部分的死代码（RepoRowView.swift 全文 -245 行）。Trending/Weekly 相关清理在后端 envelope 上线、6.2/6.3 完成后再做单独 commit，避免现在删了又复活。360/360 全绿。
 
 - [ ] **7.3** 替换 Activity / Weekly 自写刷新动画为 `SyncIconButton`
   - 通过 `RepoDetailScaffold.onRefresh` 入参传入
